Module proj17_iot_server::endpoints::devices_endpoints

Endpoints that deal rfid devices.

Statics

static_rocket_route_info_for_get_tracker
Rocket code generated static route info.

static_rocket_route_info_for_register
Rocket code generated static route info.

static_rocket_route_info_for_register_json
Rocket code generated static route info.

static_rocket_route_info_for_unregister
Rocket code generated static route info.

static_rocket_route_info_for_unregister_json
Rocket code generated static route info.

Functions

get_tracker
Responds with the receiver if any a tracker is registered at.

register
Registers a specified tracker from a specified receiver, granted both exist.

register_json
Registers a specified tracker from a specified receiver, granted both exist, where the arguments are given as request body.

rocket_route_fn_get_tracker
Rocket code generated wrapping route function.

rocket_route_fn_register
Rocket code generated wrapping route function.

rocket_route_fn_register_json
Rocket code generated wrapping route function.

rocket_route_fn_unregister
Rocket code generated wrapping route function.

rocket_route_fn_unregister_json
Rocket code generated wrapping route function.

unregister
Unregisters a specified tracker from a specified receiver. If the tracker is not currently registered to this receiver, this call will have no effect.

unregister_json
Unregisters a specified tracker from a specified receiver. If the tracker is not currently registered to this receiver, this call will have no effect.