Function proj17_iot_server::endpoints::devices_endpoints::unregister_json

pub fn unregister_json(
    body: Json<RegisterBody>
) -> Result<JsonValue, Option<JsonValue>>

Unregisters a specified tracker from a specified receiver. If the tracker is not currently registered to this receiver, this call will have no effect.

Responds with:

200: if the tracker and receiver exist.

404: if either the receiver or tracker does not exist

This is an API endpoint mapped to

/unregister [POST]

Arguments

Post body (json):

{ loc: <receiver_id>, tag: <tracker_id> }