Module proj17_iot_server::endpoints::videos_endpoints

Endpoint that deal with videos.

Statics

static_rocket_route_info_for_get_video
Rocket code generated static route info.

static_rocket_route_info_for_register_view
Rocket code generated static route info.

Functions

bad_request_builder
Utility function

get_video
Get the most appropriate video to play on the screen of specified id Appropriateness depends on the trackers currently registered to the reciver, and their interests

register_view
Registers a video view, granted the display_id, video_id, order_id, all exist. If all arguments are valid, a new played_advertisement row will be created in the database, and credits removed from the order_id. 1 credit will be withdrawn for each 8 seconds played.

rocket_route_fn_get_video
Rocket code generated wrapping route function.

rocket_route_fn_register_view
Rocket code generated wrapping route function.